Added pictures of computer teardown & labeled components
Added pictures of computer teardown & labeled components 

So in this the task at hand was to pull apart the Hardware part of my desktop computer. My main intention was to see the CPU though. Since it is not something you go about and see everyday. While I was doing the TCM Security Helpdesk course and I saw my instructor (Sir Andrew Bellini) reached the part where he had to remove the CPU from under the bracket and add new thermal paste. I was Like why not try it on the current computer I'm doing the course on.

It was well indeed a risk took and the chances of me pulling it apart on the first try and seeing the components then putting it back together and getting it back to working perfectly was pretty low. But I held onto believing in myself that I can do I and went ahead.

So the pictures above represent all the components being pulled out, and eventually I ended up cleaning the computer system on the first try and that went well. I also labelled them using sticky note paper material so each hardware component is easily identifiable.

This project helped me in such a way I can't even explain. I have done a lot of IT in Primary and Secondary school. But I was mostly faced with the theory of hardware and was always on the frontend of the computer system. And this course provided me with the necessary knowledge to pull apart computer system with even my eyes closed.